+++ src/main/kotlin/com/warus/gallery/api/model/PhotoUpdateRequest.kt
@@ -1,7 +1,7 @@
 package com.warus.gallery.api.model
 
 data class PhotoUpdateRequest(
-   val height: Int,
+   val height: Double,
    val material: String,
    val color: String,
    val type: String
